refactor: move template block data into hook [CU-8693jhuuw] (#708)
* move template block data into hook

* reorder imports. add type-only imports

* reorder imports. add type-only imports. fix initial state.

* avoid mutation. use early return.
